seguimiento malo

+++ Siscan_Vc_BLL/Service/InterfacesService/ISeguimientoService.cs
@@ -12,7 +12,7 @@ namespace Siscan_Vc_BLL.Service.InterfacesService
         Task<bool> Insert(SeguimientoInstructorAprendiz model);
         Task<bool> Update(SeguimientoInstructorAprendiz model);
         Task<bool> Delete(int id);
-        Task<SeguimientoInstructorAprendiz> GetForId(int id);
+        Task<SeguimientoInstructorAprendiz> GetForId(long id);
         Task<IQueryable<SeguimientoInstructorAprendiz>> GetAll();
         Task<SeguimientoInstructorAprendiz> GetForNumDocAprdz(string numDoc);
     }
